20130529 Spring Autowire Primitive/src/main/java/SpringAutowirePrimitive/controller/HomeController.java
package SpringAutowirePrimitive.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.servlet.ModelAndView;

import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@Controller
public class HomeController {
    @Autowired
    private Integer tryInjectInteger;

    @RequestMapping(value = "/")
    public ModelAndView test(HttpServletResponse response) throws IOException {
        return new ModelAndView("home");
    }

    public Integer getTryInjectInteger() {
        return tryInjectInteger;
    }
}
